ogc-server-statistics: replace AllTests profile by proper integration tests
The AllTests profile required to have georchestra/database running
at localhost:5432 in order to execute the tests that were actually
integration tests, hence they were never actually run.

This patch follows suite with the console module integration
test approach refactoring those tests to the src/it folder,
and uses the `it` maven profile and the io.fabric8:docker-maven-plugin
to start and stop the database container when the integration
tests are run, using dynamic port mapping to avoid conflicts
with an actual database running on localhost, and making the
tests being run by travis-ci.